重構 & 新增事件排程工具
1. 新增 dbEventSchedule collection 儲存事件觸發時間點
2. 新增 eventScheduler 輔助處理事件回呼與重覆事件的自動排程
3. 釋股與參考價更新的 counter 以新的事件排程機制取代
4. 將經理人選舉與亂鬥攻擊列表產生的流程分開
5. 用新的事件排程機制觸發經理人選舉與亂鬥攻擊列表產生
6. 將先前大部分遺留在 server/ 底下的東西拆分至 server/functions 內
7. 廢棄 counterManager
8. 經理人選舉使用 underscord 的 max 來掃描第一位最高票當選者，取代排序後取最大再找第一個的程序
