+++ server/functions/arena/startArenaFight.js
@@ -1,98 +1,23 @@
-'use strict';
 import { _ } from 'meteor/underscore';
 import { Meteor } from 'meteor/meteor';
+
 import { dbArena } from '/db/dbArena';
 import { dbArenaFighters, MAX_MANNER_SIZE, getAttributeNumber } from '/db/dbArenaFighters';
 import { dbArenaLog } from '/db/dbArenaLog';
 import { dbCompanies } from '/db/dbCompanies';
 import { dbLog } from '/db/dbLog';
 import { debug } from '/server/imports/utils/debug';
+import { getCurrentArena } from './getCurrentArena';
+import { removeUnqualifiedArenaFighters } from './removeUnqualifiedArenaFighters';
 
-// 移除未達成報名門檻的參賽者
-export function removeUnqualifiedArenaFighters(arenaData) {
-  const { _id: arenaId } = arenaData;
-  const { arenaMinInvestedAmount } = Meteor.settings.public;
-
-  const removedArenaFighterIdList = [];
-  const userRefundMap = {};
-  const logDataList = [];
-
-  dbArenaFighters
-    .aggregate([ {
-      $match: { arenaId }
-    }, {
-      $project: {
-        companyId: 1,
-        investors: 1,
-        totalInvestedAmount: { $sum: '$investors.amount' }
-      }
-    }, {
-      $match: {
-        totalInvestedAmount: { $lt: arenaMinInvestedAmount }
-      }
-    } ])
-    .forEach(({ _id: arenaFighterId, companyId, investors }) => {
-      const logCreatedAt = new Date();
-
-      logDataList.push({
-        logType: '亂鬥失格',
-        companyId,
-        createdAt: logCreatedAt
-      });
-      removedArenaFighterIdList.push(arenaFighterId);
-
-      _.each(investors, ({ userId, amount }, i) => {
-        logDataList.push({
-          logType: '亂鬥退款',
-          companyId,
-          userId: [userId],
-          data: { refund: amount },
-          createdAt: new Date(logCreatedAt.getTime() + i + 1)
-        });
-        userRefundMap[userId] = (userRefundMap[userId] || 0) + amount;
-      });
-    });
-
-  dbArenaFighters.remove({ _id: { $in: removedArenaFighterIdList }});
-
-  if (! _.isEmpty(userRefundMap)) {
-    const usersBulk = Meteor.users.rawCollection().initializeUnorderedBulkOp();
-    _.pairs(userRefundMap).forEach(([userId, refund]) => {
-      usersBulk
-        .find({ _id: userId })
-        .updateOne({ $inc: { 'profile.money': refund }});
-    });
-    Meteor.wrapAsync(usersBulk.execute).call(usersBulk);
-  }
-
-  if (logDataList.length > 0) {
-    const logBulk = dbLog.rawCollection().initializeUnorderedBulkOp();
-    logDataList.forEach((logData) => {
-      logBulk.insert(logData);
-    });
-    Meteor.wrapAsync(logBulk.execute).call(logBulk);
-  }
-
-  // 移除（剩下參賽者的）投資人資訊
-  dbArenaFighters.update({ arenaId }, { $unset: { investors: 1 } });
-}
-
+// 開始進行本次亂鬥大賽
 export function startArenaFight() {
   console.log('start arena fight...');
-  const lastArenaData = dbArena.findOne({}, {
-    sort: {
-      beginDate: -1
-    },
-    fields: {
-      _id: 1,
-      shuffledFighterCompanyIdList: 1
-    }
-  });
 
   // 正式開賽前，先移除未達成報名門檻的參賽者
-  removeUnqualifiedArenaFighters(lastArenaData);
+  removeUnqualifiedArenaFighters();
 
-  const arenaId = lastArenaData._id;
+  const { _id: arenaId, shuffledFighterCompanyIdList } = getCurrentArena();
 
   // 取得所有參賽者的數值資料與統計總投入金額
   let allFighterTotalInvest = 0;
@@ -143,8 +68,6 @@ export function startArenaFight() {
   }
 
   debug.log('startArenaFight', fighterListBySequence);
-
-  const shuffledFighterCompanyIdList = lastArenaData.shuffledFighterCompanyIdList;
 
   //輸家companyId陣列，依倒下的順序排列
   const loser = [];
@@ -281,4 +204,3 @@ export function startArenaFight() {
     }
   });
 }
-export default startArenaFight;
