+++ server/functions/company/checkChairman.js
@@ -1,34 +1,8 @@
 import { Meteor } from 'meteor/meteor';
 
-import { countdownManager } from '/server/imports/utils/countdownManager';
 import { dbCompanies } from '/db/dbCompanies';
 import { dbDirectors } from '/db/dbDirectors';
 import { debug } from '/server/imports/utils/debug';
-
-function generateCheckChairmanCounter() {
-  return Meteor.settings.public.checkChairmanCounter;
-}
-
-export function countDownCheckChairman() {
-  debug.log('countDownCheckChairman');
-
-  const counterKey = 'checkChairmanCounter';
-
-  if (! countdownManager.isInitialized(counterKey)) {
-    countdownManager.set(counterKey, generateCheckChairmanCounter());
-  }
-
-  countdownManager.countDown(counterKey);
-
-  if (! countdownManager.isZeroReached(counterKey)) {
-    return;
-  }
-
-  const nextCounter = generateCheckChairmanCounter();
-  countdownManager.set(counterKey, nextCounter);
-
-  checkChairman();
-}
 
 // 對全市場進行董事長檢查
 export function checkChairman() {
