+++ server/paySalaryAndCheckTax.js
@@ -16,41 +16,27 @@ const {salaryPerPay} = Meteor.settings.public;
 export function paySalaryAndCheckTax() {
   debug.log('paySalary');
   const todayBeginTime = new Date().setHours(0, 0, 0, 0);
-  const lastPayTime = dbVariables.get('lastPayTime');
-  if (! lastPayTime || lastPayTime.setHours(0, 0, 0, 0) !== todayBeginTime) {
-    if (Date.now() - lastPayTime <= 86400000) {
-      console.error('paySalaryAndCheckTax error!', Date.now(), todayBeginTime, lastPayTime, lastPayTime.setHours(0, 0, 0, 0));
-
-      return false;
-    }
-    const thisPayTime = new Date();
-    dbVariables.set('lastPayTime', thisPayTime);
-    resourceManager.request('paySalaryAndCheckTax', ['season'], (release) => {
-      paySystemSalary(thisPayTime);
-      paySalaryAndGenerateProfit(thisPayTime);
-      checkTax(todayBeginTime);
-      release();
-    });
-  }
+  const thisPayTime = new Date();
+  dbVariables.set('lastPayTime', thisPayTime);
+  resourceManager.request('paySalaryAndCheckTax', ['season'], (release) => {
+    paySystemSalary(thisPayTime);
+    paySalaryAndGenerateProfit(thisPayTime);
+    checkTax(todayBeginTime);
+    release();
+  });
 }
 
 function paySystemSalary(thisPayTime) {
   console.info(thisPayTime.toLocaleString() + ': paySystemSalary');
-  Meteor.users.update(
-    {
-      createdAt: {
-        $lte: thisPayTime
-      }
-    },
-    {
-      $inc: {
-        'profile.money': salaryPerPay
-      }
-    },
-    {
-      multi: true
-    }
-  );
+
+  Meteor.users.update({
+    createdAt: { $lte: thisPayTime }
+  }, {
+    $inc: { 'profile.money': salaryPerPay }
+  }, {
+    multi: true
+  });
+
   dbLog.insert({
     logType: '發薪紀錄',
     userId: ['!all'],
